+++ include/renderers.hpp
@@ -38,7 +38,6 @@ public:
   //-----------Constructors-----------
   Renderer(std::shared_ptr<World> world, Color background_color = Color()) : world(world), background_color(background_color) {};
 
-
   //------------Methods-----------
   virtual Color operator()(Ray ray) const = 0;
 };
@@ -89,12 +88,11 @@ public:
     if (!hit) {
       return background_color;
     }
-    // Return the color of the closest object hit (if any), both the brdf pigment and the emitted radiance 
+    // Return the color of the closest object hit (if any), both the brdf pigment and the emitted radiance
     return (*(hit->shape->material->brdf->pigment))(hit->surface_point) +
            (*(hit->shape->material->emitted_radiance))(hit->surface_point);
   };
 };
-
 
 // ------------------------------------------------------------------------------------------------------------
 // --------------- POINT LIGHT TRACER -------------------------------------------------
@@ -133,8 +131,10 @@ public:
       hit_material = hit->shape->material;
       brdf = hit_material->brdf;
 
-      auto specular = std::dynamic_pointer_cast<SpecularBRDF>(brdf); // returns nullptr if brdf is not a pointer to an object of the derived class SpecularBRDF
-      // 3. In case you hit an object with SpecularBRDF, scatter a new Ray from the hit point in the direction given by the reflection law; otherwise go ahead with color evaluation
+      auto specular = std::dynamic_pointer_cast<SpecularBRDF>(
+          brdf); // returns nullptr if brdf is not a pointer to an object of the derived class SpecularBRDF
+      // 3. In case you hit an object with SpecularBRDF, scatter a new Ray from the hit point in the direction given by the
+      // reflection law; otherwise go ahead with color evaluation
       if (!specular) {
         break;
       }
@@ -154,7 +154,7 @@ public:
 
       if (in_dir.has_value()) {
         float distance = in_dir->norm();
-        float distance_factor = (source->emission_radius > 0.f) ? std::pow((source->emission_radius / distance), 2) : 1.f;
+        float distance_factor = (source->emission_radius > 0.f) ? std::powf((source->emission_radius / distance), 2) : 1.f;
         // angle between normal at hitting point and incoming direction (from light source)
         float cos_theta = (-1.f / distance) * (*in_dir) * (1.f / hit->normal.norm()) *
                           (hit->normal); // QUESTION why the minus sign? and why we multiply by costheta?
@@ -169,7 +169,6 @@ public:
 // ------------------------------------------------------------------------------------------------------------
 // --------------- PATH TRACER -------------------------------------------------
 // ------------------------------------------------------------------------------------------------------------
-
 
 /// @brief functor performing the path tracing algorithm
 /// @details importance sampling in MC integration based on the scatter_ray method of the BRDF
